+++ src/main/java/sk/kas/mirda/Cameras/repository/CameraRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import sk.kas.mirda.Cameras.model.Camera;
 
 @Repository
-public interface CameraRepository extends JpaRepository<Camera, Integer> {
+public interface CameraRepository extends JpaRepository<Camera, Long> {
 }
